fix(kraeplin): add condition in down section for adding active column id

+++ database/migrations/tenant/2019_09_17_095050_add_active_column_id_to_psychology_kraeplins_table.php
@@ -31,8 +31,10 @@ class AddActiveColumnIdToPsychologyKraeplinsTable extends Migration
      */
     public function down()
     {
-        Schema::table('psychology_kraeplin', function (Blueprint $table) {
-            $table->dropColumn('active_column_id');
-        });
+        if(Schema::hasColumn('psychology_kraeplins', 'active_column_id')) {
+            Schema::table('psychology_kraeplin', function (Blueprint $table) {
+                $table->dropColumn('active_column_id');
+            });
+        }
     }
 }
